+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Release 0.2.0
+* **The release breaks backwards compatibility for `prependWild`, `appendWild`, and automatically adding `<td class="icon-filter"><i>` element.**
+* Complete re-write
+* Added `isMatch(cell, query)` option
+* Added `editableSelector` option
+* Added `editableOptions` option
+
 ## 2013-07-04 - Release 0.1.3
 * Escape user filter to prevent it from being interpreted as regex - thanks Reese purpletreefactory@1466e15b8977a7f177fd7cf12e01989566fec04b